cleanup(core): extract shared import e2e test helpers
