+++ e2e/nx/src/import-ai-agent.test.ts
@@ -9,44 +9,14 @@ import {
   updateFile,
   updateJson,
 } from '@nx/e2e-utils';
-import { writeFileSync, mkdirSync, rmdirSync } from 'fs';
-import { execSync } from 'node:child_process';
+import { rmdirSync } from 'fs';
 import { join } from 'path';
-
-/**
- * Parse NDJSON output from agent-mode nx commands.
- * Splits stdout on newlines, parses each JSON line,
- * filters out USER_NEXT_STEPS plain text section.
- */
-function parseNdjsonOutput(stdout: string): Record<string, any>[] {
-  return stdout
-    .split('\n')
-    .filter((line) => {
-      const trimmed = line.trim();
-      if (!trimmed) return false;
-      // Skip the USER_NEXT_STEPS plain text section
-      if (trimmed.startsWith('---') || trimmed.startsWith('[DISPLAY]'))
-        return false;
-      // Only try to parse lines starting with {
-      if (!trimmed.startsWith('{')) return false;
-      return true;
-    })
-    .map((line) => {
-      try {
-        return JSON.parse(line.trim());
-      } catch {
-        return null;
-      }
-    })
-    .filter(Boolean);
-}
-
-function findMessage(
-  messages: Record<string, any>[],
-  predicate: (msg: Record<string, any>) => boolean
-): Record<string, any> | undefined {
-  return messages.find(predicate);
-}
+import {
+  createSimpleRepo,
+  createMultiPackageRepo,
+  parseNdjsonOutput,
+  findMessage,
+} from './import-utils';
 
 const agentEnv = { CLAUDECODE: '1' };
 
@@ -80,62 +50,9 @@ describe('Nx Import - AI Agent Mode', () => {
 
   afterAll(() => cleanupProject());
 
-  /**
-   * Helper: create a simple git repo with a README
-   */
-  function createSimpleRepo(name: string): string {
-    const repoPath = join(tempImportE2ERoot, name);
-    mkdirSync(repoPath, { recursive: true });
-    writeFileSync(join(repoPath, 'README.md'), `# ${name}`);
-    writeFileSync(
-      join(repoPath, 'package.json'),
-      JSON.stringify({ name, version: '1.0.0' }, null, 2)
-    );
-    execSync(`git init && git add . && git commit -m "initial commit"`, {
-      cwd: repoPath,
-    });
-    try {
-      execSync(`git checkout -b main`, { cwd: repoPath });
-    } catch {}
-    return repoPath;
-  }
-
-  /**
-   * Helper: create a repo with two packages
-   */
-  function createMultiPackageRepo(): string {
-    const repoPath = join(tempImportE2ERoot, 'multi-pkg');
-    mkdirSync(repoPath, { recursive: true });
-    writeFileSync(join(repoPath, 'README.md'), '# Repo');
-    execSync(`git init && git add . && git commit -m "initial"`, {
-      cwd: repoPath,
-    });
-    try {
-      execSync(`git checkout -b main`, { cwd: repoPath });
-    } catch {}
-
-    mkdirSync(join(repoPath, 'packages/a'), { recursive: true });
-    writeFileSync(join(repoPath, 'packages/a/README.md'), '# A');
-    writeFileSync(
-      join(repoPath, 'packages/a/package.json'),
-      JSON.stringify({ name: 'pkg-a', version: '1.0.0' }, null, 2)
-    );
-    execSync(`git add . && git commit -m "add package a"`, { cwd: repoPath });
-
-    mkdirSync(join(repoPath, 'packages/b'), { recursive: true });
-    writeFileSync(join(repoPath, 'packages/b/README.md'), '# B');
-    writeFileSync(
-      join(repoPath, 'packages/b/package.json'),
-      JSON.stringify({ name: 'pkg-b', version: '1.0.0' }, null, 2)
-    );
-    execSync(`git add . && git commit -m "add package b"`, { cwd: repoPath });
-
-    return repoPath;
-  }
-
   describe('needs_input for missing arguments', () => {
     it('should return needs_input when destination is missing', () => {
-      const repoPath = createSimpleRepo('needs-input-test');
+      const repoPath = createSimpleRepo(tempImportE2ERoot, 'needs-input-test');
       const output = runCLI(`import ${repoPath} --ref main`, {
         silenceError: true,
         env: agentEnv,
@@ -172,14 +89,13 @@ describe('Nx Import - AI Agent Mode', () => {
 
   describe('full import with agent mode', () => {
     it('should complete successfully with --plugins=skip', () => {
-      const repoPath = createSimpleRepo('plugins-skip-test');
+      const repoPath = createSimpleRepo(tempImportE2ERoot, 'plugins-skip-test');
       const output = runCLI(
         `import ${repoPath} projects/skip-test --ref main --source . --plugins skip`,
         { verbose: true, env: agentEnv }
       );
       const messages = parseNdjsonOutput(output);
 
-      // Check progress stages
       expect(
         findMessage(messages, (m) => m.stage === 'starting')
       ).toBeDefined();
@@ -205,7 +121,7 @@ describe('Nx Import - AI Agent Mode', () => {
 
   describe('import subdirectory', () => {
     it('should import a subdirectory with correct paths in output', () => {
-      const repoPath = createMultiPackageRepo();
+      const repoPath = createMultiPackageRepo(tempImportE2ERoot);
       const output = runCLI(
         `import ${repoPath} projects/pkg-a --ref main --source packages/a --plugins skip`,
         { verbose: true, env: agentEnv }
@@ -225,16 +141,17 @@ describe('Nx Import - AI Agent Mode', () => {
 
   describe('error cases', () => {
     it('should return structured error for non-empty destination', () => {
-      const repoPath = createSimpleRepo('error-nonempty-test');
+      const repoPath = createSimpleRepo(
+        tempImportE2ERoot,
+        'error-nonempty-test'
+      );
 
-      // First import to populate destination
       runCLI(
         `import ${repoPath} projects/error-dest --ref main --source . --plugins skip`,
         { verbose: true, env: agentEnv }
       );
       runCommand(`git add . && git commit -am "first import"`);
 
-      // Second import to same destination should fail
       const output = runCLI(
         `import ${repoPath} projects/error-dest --ref main --source . --plugins skip`,
         { silenceError: true, env: agentEnv }
